feat(resolvers): add bootstrap-vue-next resolver
chore: move bootstrap-vue-3 resolver, and add note about deprecation
