+++ src/core/resolvers/index.ts
@@ -19,5 +19,4 @@ export * from './arco'
 export * from './tdesign'
 export * from './layui-vue'
 export * from './bootstrap-vue'
-export * from './bootstrap-vue-3'
 export * from './ionic'
